+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # CHANGELOG
 
-## 4.2.18
+## 4.3.0
 
 - tidy3d improvements:
     - update to version 1.1.1
-- remove casting port orientation to int
+- change port angle type annotation from int to float
 
 ## [4.2.17](https://github.com/gdsfactory/gdsfactory/pull/210)